requirements.txt
# Core
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
pymongo
motor>=3.3.0

# Data Validation
pydantic[email]>=1.10.7
email-validator>=2.0.0

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Kafka
kafka-python>=2.0.2

# Development
pytest>=7.3.1
httpx>=0.24.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0